+++ src/commands/index.ts
@@ -164,6 +164,7 @@ import { econnreset } from './support/econnreset';
 import { pr } from './utils/pr';
 import { gsxIntegration } from './aircraft/gsxIntegration';
 import { supportGuide } from './support/supportGuide';
+import { simbrief } from './utils/simbrief';
 
 const commands: BaseCommandDefinition[] = [
     typeCommand,
@@ -330,6 +331,7 @@ const commands: BaseCommandDefinition[] = [
     pr,
     gsxIntegration,
     supportGuide,
+    simbrief
 ];
 
 const commandsObject: { [k: string]: BaseCommandDefinition } = {};